add revert.models.Model class for writer/optimize
