+++ revert/models/__init__.py
@@ -1,5 +1,6 @@
 from .neural_decomposition import ND
 from .barlow_twins  import BarlowTwins, cross_correlation, norm2
 from .conv          import ConvNet
+from .model         import Model
 from .geometry      import tsne, mdse, pca
 from .kmeans        import KMeans
