+++ revert/models/barlow_twins.py
@@ -1,6 +1,8 @@
 import torch
 import torch.nn as nn
 
+from .model import Model
+
 def norm2 (t, dim=None):
     """ L2-norm on specified dimensions """
     return torch.sqrt((t ** 2).sum(dim)) 
@@ -11,7 +13,8 @@ def cross_correlation (ya, yb):
     yab = ya[:,:,None] @ yb[:,None,:]
     return yab.sum(dim=[0]) / (norm2(ya, [0]) * norm2(yb, [0]))
 
-class BarlowTwins (nn.Module):
+
+class BarlowTwins (Model):
 
     def __init__(self, model, diag=2):
         """ Create twins from a model. """
@@ -36,71 +39,5 @@ class BarlowTwins (nn.Module):
         return torch.sum(((C - I) * loss_mask) ** 2) / (n_out ** 2)
 
     def cross_corr (self, x):
-        """ Cross correlation matrix of twin outputs. l"""
+        """ Cross correlation matrix of twin outputs. """
         return cross_correlation(*self(x))
-
-    def loss_on (self, x):
-        """ Barlow twin loss on input """
-        return self.loss(self.forward(x))
-    
-    def optimize (self, xs, optimizer, scheduler=None, epochs=1, w=None):
-        """ Fit on a N_it x 2 x N_batch x N tensor. """
-        N_it = xs.shape[0]
-        for e in range(epochs):
-            for nit, x in enumerate(xs): 
-                optimizer.zero_grad()
-                loss = self.loss_on(x)
-                loss.backward()
-                optimizer.step()
-                if w:
-                    self.write(w, loss.detach(), nit + e * N_it) 
-            if scheduler:
-                scheduler.step()
-        return self
-
-    def fit (self, x, lr=1e-2, br=1e-3, n_batch=128, w="Loss/fit"):
-        """ Fit on a 2 x N_samples x N tensor. """
-        n_it = x.shape[1] // n_batch
-        print(f"Fitting on {n_it} * {n_batch} pairs...")
-        for s in range(n_it):
-            y = self.forward(x[:,s:s + n_batch])
-            loss = self.loss(y)
-            loss.backward()
-            if w: 
-                self.write(w, loss, nit=s)
-            with torch.no_grad(): 
-                for p in self.parameters(): 
-                    p -= p.grad * lr
-                    p -= br * torch.randn(p.shape, device=p.device)
-                self.zero_grad()
-        return self
-
-    def loop (self, xs, lr=1e-2, br=1e-3, n_batch=128, w="Loss/fit", epochs=1):
-        lr = lr if isinstance(lr, list) else [lr] * epochs
-        br = br if isinstance(br, list) else [br] * epochs
-        lr = lr + [lr[-1]] * (epochs - len(lr))
-        br = br + [br[-1]] * (epochs - len(br))
-        for e in range(epochs):
-            ys   = self(xs)
-            l0   = self.loss(ys)
-            print(f"\nLoss {e}: {float(l0):.4f}")
-            self.fit(xs, lr[e], br[e], n_batch, w=f'{w}{e}')
-        ys   = self(xs)
-        l1   = self.loss(ys)
-        print(f"\nLoss {e}: {float(l1):.4f}")
-        return self
-
-
-
-    def write(self, name, data, nit):
-        if self.writer:
-            self.writer.add_scalar(name, data, global_step=nit)
-
-
-
-
-
-
-        
-         
-
